Marginalization Strategies

Direct PWS

PWS.DirectMCEstimate — Type

DirectMCEstimate(M::Int)

This is the simplest marginalization strategy, estimating the marginal path probabilities using a direct Monte Carlo simulation.

M specifies the number of samples to be used for the Monte Carlo average. Larger M improves the accuracy of the marginalization integral, but increases the computational cost.

Mathematical Description

The marginal probability

\[\mathrm{P}[\bm{x}] = \int\mathrm{d}\bm{s} \mathrm{P}[\bm{x}|\bm{s}] \mathrm{P}[\bm{s}]\]

can be computed via a Monte Carlo estimate by sampling M trajectories from $\mathrm{P}[\bm{s}]$ and taking the average of the likelihoods:

\[\mathrm{P}[\bm{x}] = \langle \mathrm{P}[\bm{x}|\bm{s}] \rangle_{\mathrm{P}[\bm{s}]}\,.\]

RR-PWS

PWS.SMCEstimate — Type

SMCEstimate(num_particles::Int)

Compute the marginal trajectory probability using a Sequential Monte Carlo (SMC) algorithm, more specifically using a particle filter.

In a particle filter, num_particles trajectories are propagated in parallel. At regular intervals, the current set of parallel trajectories undergoes a resampling step where some of the trajectories get eliminated and others duplicated, depending on their accumulated likelihood. This resampling ensures that we don't sample trajectories that contribute only very little to the marginalization integral.

Due to the periodic resampling of the trajectories, this method works much better than the DirectMCEstimate for long trajectories.

TI-PWS

PWS.TIEstimate — Type

TIEstimate(burn_in::Int, integration_nodes::Int, num_samples::Int)

Compute marginalization integrals using thermodynamic integration.