Systems

Example Systems included in PathWeightSampling.jl

PathWeightSampling.gene_expression_system — Function

gene_expression_system(; mean_s=50, mean_x=mean_s, corr_time_s=1.0, corr_time_x=0.1, u0=SA[mean_s, mean_x], dtimes=0:0.1:2.0)

Creates a system for a very simple model of gene expression.

Model Description

This model consists of four reactions:

                κ
reaction 1:  ∅ ---> S

                λ
reaction 2:  S ---> ∅

                ρ
reaction 3:  S ---> S + X

                μ
reaction 4:  X ---> ∅

The first two reactions specify the evolution of the input signal S, and last two reactions specify the evolution of the output X. Thus, both the input and output signal are modeled as a simple birth-death process, however the birth rate of X increases with higher copy numbers of S.

Examples

The values of the reaction rates can be specified directly as follows:

using PathWeightSampling
system = PathWeightSampling.gene_expression_system(kappa = 10.0, lambda = 0.1, rho = 1.0, mu = 1.0)

# output

SimpleSystem with 4 reactions
Input variables: S(t)
Output variables: X(t)
Initial condition:
    S(t) = 50
    X(t) = 50
Parameters:
    κ = 10.0
    λ = 0.1
    ρ = 1.0
    μ = 1.0

Alternatively, the reaction rates can be specified indirectly through the following arguments:

mean_s: The average copy number of S when the system has relaxed to steady state.

mean_x: The average copy number of X when the system has relaxed to steady state.

corr_time_s: The input signal correlation time. The shorter this time is, the faster the fluctuations in the input signal.

corr_time_x: The output signal correlation time. This sets the timescale of output fluctuations.

using PathWeightSampling
system = PathWeightSampling.gene_expression_system(mean_s=25, mean_x=50, corr_time_s=1.0, corr_time_x=0.75)

# output

Missing docstring.

Missing docstring for PathWeightSampling.cooperative_chemotaxis_system. Check Documenter's build log for details.

Create a New System

To create a new system, you can make either an object of type SimpleSystem or ComplexSystem from scratch. The difference between the two is that a ComplexSystem represents a model that includes latent variables which need to be integrated out to compute the mutual information. A SimpleSystem represents a model without any latent variables.

Both types of systems are constructed using Catalyst's representation of a reaction network (i.e. from objects of type ReactionSystem). Those reaction systems can be constructed using the API from Catalyst.jl.

Missing docstring.

Missing docstring for PathWeightSampling.SimpleSystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PathWeightSampling.ComplexSystem. Check Documenter's build log for details.